requirements.txt
Flask==2.3.2
psycopg2==2.9.7
requests==2.31.0
celery==5.3.1
python-dotenv~=1.0.0